Backend_Catalog_Microservice: define Pagination (Result) and CatalogSpecParams (Input) usomg specification pattern

+++ Ecommerce/Services/Catalog/Specifications/Pagination.cs
@@ -2,12 +2,12 @@
 {
     public class Pagination<T> where T : class
     {
-        public Pagination()
+        public Pagination() 
         {
             
         }
 
-        public Pagination(int pageIndex, int pageSize, int count, IReadOnlyCollection<T> data)
+        public Pagination(int pageIndex, int pageSize, long count, IReadOnlyCollection<T> data)
         {
             PageIndex = pageIndex;
             PageSize = pageSize;
@@ -17,7 +17,7 @@
 
         public int PageIndex { get; set; }
         public int PageSize { get; set; }
-        public int Count { get; set; }
+        public long Count { get; set; }
         public IReadOnlyCollection<T> Data { get; set; }
     }
 }
